Getting the Id's in shape
It's been hard AF

+++ src/dirserver/ServerRegistry.java
@@ -7,9 +7,7 @@ import dirserver.Registries;
  */
 public class ServerRegistry extends Registries {
 
-
-
-    ServerRegistry(String n,int udp,int tcp,int entry){
+    ServerRegistry(String n,int udp,int tcp,long entry){
         name=n;
         udpPort=udp;
         tcpPort=tcp;
